VIT-7771: Fix deprioritized resources not making forward progress in some environments (#253)
React Native does not appear to use UIScene.

As a result, the app lifecycle callouts appear to behave quite differently. The `appWillEnterForeground` callout we expected after `appDidFinishLaunching` does not happen.

This causes our `AppStateTracker` to be stuck in `state = appLaunching` instead of progressing to `state = foreground`. This in turn has a knock-on effect on deprioritised resources, since re-attempting deprioritised resources is skipped unless the app is in foreground.

This PR:

1. schedules a `UIApplication.applicationState` check one runloop iteration after `appDidFinishLaunching`.

    * This ensures that `AppStateTracker` will transition to `state = foreground` in both UIScene-disabled and UIScene-enabled apps.

2. Expand the scope of resource rescue of `scheduleUnnotifiedResourceRescue` to include resources stuck in `started` and `deprioritized` states.